+++ reproc/include/reproc/reproc.h
@@ -120,6 +120,17 @@ typedef struct reproc_options {
   reproc_stop_actions stop_actions;
 } reproc_options;
 
+/*! Used by `reproc_drain` to provide data to the caller. Each time data is
+read, `function` is called with `context`. See `reproc_drain` and the `drain`
+example for more information .*/
+typedef struct reproc_sink {
+  bool (*function)(REPROC_STREAM stream,
+                   const uint8_t *buffer,
+                   size_t size,
+                   void *context);
+  void *context;
+} reproc_sink;
+
 /*! Allocate a new `reproc_t` instance on the heap. */
 REPROC_EXPORT reproc_t *reproc_new(void);
 
@@ -171,11 +182,16 @@ REPROC_EXPORT int reproc_read(reproc_t *process,
                               size_t size);
 
 /*!
-Calls `reproc_read` on `stream` until `reproc_read` returns an error or `sink`
-returns false. `sink` receives the output after each read, along with `context`.
+Calls `reproc_read` on `stream` until `reproc_read` returns an error or one of
+the sinks returns false. The `out` and `err` sinks receive the output from
+stdout and stderr respectively. The same sink may be passed to both `out` and
+`err`.
 
-`reproc_drain` always starts by calling `sink` once with an empty buffer and
-`stream` set to `REPROC_STREAM_IN` to give the sink the chance to process all
+If `out` or `err` are `NULL`, all output on the corresponding stream is
+discarded.
+
+`reproc_drain` always starts by calling both sinks once with an empty buffer and
+`stream` set to `REPROC_STREAM_IN` to give each sink the chance to process all
 output from the previous call to `reproc_drain` one by one.
 
 Note that his function returns 0 instead of `REPROC_EPIPE` when both output
@@ -183,12 +199,8 @@ streams of the child process are closed.
 
 For examples of sinks, see `sink.h`.
 */
-REPROC_EXPORT int reproc_drain(reproc_t *process,
-                               bool (*sink)(REPROC_STREAM stream,
-                                            const uint8_t *buffer,
-                                            size_t size,
-                                            void *context),
-                               void *context);
+REPROC_EXPORT int
+reproc_drain(reproc_t *process, reproc_sink *out, reproc_sink *err);
 
 /*!
 Writes `size` bytes from `buffer` to the standard input (stdin) of the child