+++ reproc/test/io.c
@@ -23,7 +23,8 @@ static int io(const char *mode, const char *input, const char *expected)
   assert(r == 0);
 
   char *output = NULL;
-  r = reproc_drain(process, reproc_sink_string, &output);
+  reproc_sink sink = { reproc_sink_string, &output };
+  r = reproc_drain(process, &sink, &sink);
   assert(r == 0);
   assert(output != NULL);