+++ reprocxx/examples/cmake-help.cpp
@@ -38,12 +38,13 @@ int main()
 
   // `process::drain` reads from the stdout and stderr streams of the child
   // process until both are closed or an error occurs. Providing it with a
-  // string sink makes it store all output in the string(s) passed to the string
-  // sink. Passing the same string to both the `out` and `err` arguments of
-  // `sink::string` causes the stdout and stderr output to get stored in the
-  // same string.
+  // string sink for a specific stream makes it store all output of that stream
+  // in the string passed to the string sink. Passing the same sink to both the
+  // `out` and `err` arguments of `process::drain` causes the stdout and stderr
+  // output to get stored in the same string.
   std::string output;
-  ec = process.drain(reproc::sink::string(output, output));
+  reproc::sink::string sink(output);
+  ec = process.drain(sink, sink);
   if (ec) {
     return fail(ec);
   }